requirements.txt
segno>=1.6.0
Pillow>=10.0.0
